[FLINK-29097][sql-gateway]Moving json se/deserializers from sql-gateway-api to sql-gateway
