+++ flink-table/flink-sql-gateway/src/main/java/org/apache/flink/table/gateway/rest/message/statement/FetchResultsResponseBody.java
@@ -20,8 +20,12 @@ package org.apache.flink.table.gateway.rest.message.statement;
 
 import org.apache.flink.runtime.rest.messages.ResponseBody;
 import org.apache.flink.table.gateway.api.results.ResultSet;
+import org.apache.flink.table.gateway.rest.util.JsonResultSetDeserializer;
+import org.apache.flink.table.gateway.rest.util.JsonResultSetSerializer;
 
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonProperty;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.annotation.JsonSerialize;
 
 import javax.annotation.Nullable;
 
@@ -33,6 +37,8 @@ public class FetchResultsResponseBody implements ResponseBody {
     private static final String FIELD_NEXT_RESULT_URI = "nextResultUri";
 
     @JsonProperty(FIELD_RESULTS)
+    @JsonSerialize(using = JsonResultSetSerializer.class)
+    @JsonDeserialize(using = JsonResultSetDeserializer.class)
     private final ResultSet results;
 
     @JsonProperty(FIELD_RESULT_TYPE)
