+++ flink-table/flink-sql-gateway/src/test/java/org/apache/flink/table/gateway/rest/JsonResultSetSerDeTest.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.gateway.api.results.serde;
+package org.apache.flink.table.gateway.rest;
 
 import org.apache.flink.table.api.DataTypes;
 import org.apache.flink.table.catalog.ResolvedSchema;
@@ -24,11 +24,14 @@ import org.apache.flink.table.data.GenericRowData;
 import org.apache.flink.table.data.RowData;
 import org.apache.flink.table.data.util.DataFormatConverters;
 import org.apache.flink.table.gateway.api.results.ResultSet;
+import org.apache.flink.table.gateway.rest.util.JsonResultSetDeserializer;
+import org.apache.flink.table.gateway.rest.util.JsonResultSetSerializer;
 import org.apache.flink.table.types.DataType;
 import org.apache.flink.types.Row;
 import org.apache.flink.types.RowKind;
 
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.databind.module.SimpleModule;
 
 import org.junit.jupiter.api.Test;
 
@@ -72,7 +75,7 @@ import static org.apache.flink.table.api.DataTypes.TINYINT;
 import static org.assertj.core.api.Assertions.assertThat;
 
 /** Tests for {@link JsonResultSetSerializer} and {@link JsonResultSetDeserializer}. */
-class JsonResultSetSerializationTest {
+class JsonResultSetSerDeTest {
 
     private static final byte tinyint = 'c';
     private static final short smallint = 128;
@@ -137,13 +140,17 @@ class JsonResultSetSerializationTest {
     void seDeResultSet(List<Row> rowList, List<DataTypes.Field> fields) throws IOException {
         List<RowData> rowDataList =
                 rowList.stream()
-                        .map(JsonResultSetSerializationTest::convertToInternal)
+                        .map(JsonResultSetSerDeTest::convertToInternal)
                         .collect(Collectors.toList());
         ResolvedSchema testResolvedSchema = getTestResolvedSchema(fields);
         ResultSet testResultSet =
                 new ResultSet(ResultSet.ResultType.PAYLOAD, 0L, testResolvedSchema, rowDataList);
         // Test serialization & deserialization
         ObjectMapper objectMapper = new ObjectMapper();
+        SimpleModule resultSetModule = new SimpleModule();
+        resultSetModule.addSerializer(ResultSet.class, new JsonResultSetSerializer());
+        resultSetModule.addDeserializer(ResultSet.class, new JsonResultSetDeserializer());
+        objectMapper.registerModule(resultSetModule);
         String result = objectMapper.writeValueAsString(testResultSet);
         ResultSet resultSet = objectMapper.readValue(result, ResultSet.class);
         List<RowData> deRowDataList = resultSet.getData();
